# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SIHIYAN - SIHI Seeds AI Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock:
    get:
      operationId: listStockItems
      tags: [stock]
      summary: List all stock items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStockItem
      tags: [stock]
      summary: Create a new stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemInput"
      responses:
        "201":
          description: Created stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
...
  /stock/{id}:
    get:
      operationId: getStockItem
      tags: [stock]
      summary: Get a stock item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateStockItem
      tags: [stock]
      summary: Update a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemUpdate"
      responses:
        "200":
          description: Updated stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteStockItem
      tags: [stock]
      summary: Delete a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stock/{id}/adjust:
    post:
      operationId: adjustStockQuantity
      tags: [stock]
      summary: Adjust stock quantity (add or subtract)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustInput"
      responses:
        "200":
          description: Updated stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"

  # ─── DISPATCH MANAGEMENT ─────────────────────────────────────────────────────
...
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List all inventory records
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/warehouses:
    get:
      operationId: listWarehouses
      tags: [inventory]
      summary: List all warehouses
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [inventory]
      summary: Create a warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ANALYTICS ─────────────────────────────────────────────────────
...
  /reports/daily/{id}:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get a specific daily report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"

  # ─── OCR BILL SCANNING ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── STOCK ──────────────────────────────────────────────────────────────────
    StockItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        quantity:
          type: number
        unit:
          type: string
        minQuantity:
          type: number
        maxQuantity:
          type: ["number", "null"]
        unitPrice:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - sku
        - quantity
        - unit
        - minQuantity
        - createdAt
        - updatedAt
    StockItemInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        quantity:
          type: number
        unit:
          type: string
        minQuantity:
          type: number
        maxQuantity:
          type: ["number", "null"]
        unitPrice:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - category
        - sku
        - quantity
        - unit
        - minQuantity
    StockItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        quantity:
          type: number
        unit:
          type: string
        minQuantity:
          type: number
        maxQuantity:
          type: ["number", "null"]
        unitPrice:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    StockAdjustInput:
      type: object
      properties:
        delta:
          type: number
        reason:
          type: string
      required:
        - delta
        - reason
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        capacity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - createdAt
    WarehouseInput:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - location
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        stockItemId:
          type: integer
        warehouseId:
          type: integer
        quantity:
          type: number
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        stockItemName:
          type: ["string", "null"]
        warehouseName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - stockItemId
        - warehouseId
        - quantity
        - createdAt
        - updatedAt
    InventoryItemInput:
      type: object
      properties:
        stockItemId:
          type: integer
        warehouseId:
          type: integer
        quantity:
          type: number
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
      required:
        - stockItemId
        - warehouseId
        - quantity
    InventoryItemUpdate:
      type: object
      properties:
        quantity:
          type: number
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]

    # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
    DailyReport:
      type: object
      properties:
        id:
          type: integer
        reportDate:
          type: string
          format: date-time
        summary:
          type: string
        stockInsights:
          type: ["string", "null"]
        dispatchInsights:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        generatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reportDate
        - summary
        - generatedBy
        - createdAt

    # ─── OCR ────────────────────────────────────────────────────────────────────
...