# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SIHIYAN - SIHI Seeds AI Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyReport {
  id: number;
  reportDate: Date;
  summary: string;
  /** @nullable */
  stockInsights?: string | null;
  /** @nullable */
  dispatchInsights?: string | null;
  /** @nullable */
  recommendations?: string | null;
  generatedBy: string;
  createdAt: Date;
}